backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.6.4
pydantic-settings==2.2.1
supabase==2.4.1
twilio==8.13.0
google-generativeai==0.5.0
openai==1.23.2
newspaper3k==0.2.8
beautifulsoup4==4.12.3
lxml==5.2.1
python-dotenv==1.0.1
websockets==12.0
asyncio==3.4.3
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
